requirements.txt
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
openai>=1.0.0
unstructured>=0.10.0
beautifulsoup4>=4.12.0
pypdf>=3.0.0
markdown>=3.5.0
